PALEOsediment Reactions

Sediment transport

PALEOsediment.Sediment.SedimentTransport.ReactionSedimentTransport — Type

SedimentTransport

Sediment transport for n x 1D sediment columns.

Physical environment

A grid with n columns is created in a Domain sediment, bounded at the top by Domain oceanfloor and at the base by Domain sedimentfloor. Boundary cells at the sediment surface are therefore in subdomain sediment.oceanfloor.

The number of columns, and column area, water depth, accumulation rate, and porosity are set by oceanfloor Variables. Bioturbation and bioirrigation rates should be supplied on the sediment grid eg by PALEOsediment.Sediment.SedimentBioRates.ReactionSedimentBioRates.

Transport components and species

Each component <totalname> to be transported should be defined by a source-minus-sink flux <totalname>_sms, and one or more concentration Variables with names of form <speciesname>_conc

Solute concentration Variables are identified by attributes vphase == VP_Solute and advect == true, and are transported by diffusion, bioturbation and bioirrigation, and advection. Species-specific solute diffusivities are calculated either based on the attribute diffusivity_speciesname of the <speciesname>_conc solute Variables, which should be one of the names available from PALEOaqchem.MolecularDiffusion.create_solute_diffusivity_func, or (if diffusivity_speciesname is not set ie an empty string) by the attribute diffusivity (units cm^2 s-1).

Solid concentration Variables are identified by attributes vphase == VP_Solid and advect == true, and are transported by bioturbation and advection.

Transport fluxes are then accumulated into vectors <totalname>_sms:

If the totalnames attribute is not set on a variable <speciesname>_conc, then totalname is assumed to be the same as speciesname and transport fluxes are accumulated into <speciesname>_sms.

If the totalnames attribute is set to a Vector of strings ["mult_1*totalname_1", ... "mult_n*totalname_n"], then this is used to define the appropriate vector of [<totalname_n>_sms] fluxes (allowing multiple species concentrations with different transport properties or phases for each <totalname>_sms, and each species to contribute to multiple <totalname_n>_sms with a stoichiometry multiplier mult_n).

Boundary conditions

Oceanfloor solute fluxes should be defined in the Domain fluxOceanfloor, with names fluxOceanfloor.soluteflux_<totalname>. Input particulate fluxes should be added by the fluxOceanfloor flux coupler to the surface sediment cells by linking to Variables in the sediment.oceanfloor subdomain, ie to sediment.oceanfloor.<totalname>_sms.

Burial fluxes are the base of the sediment columns are defined in the Domain fluxOceanBurial.

Parameters

L[Float64]=0.15 (m), default_value=0.15, description="depth of sediment column"

ncellspercol[Int64]=60, default_value=60, description="number of cells per column"

f_grid[String]="linear", default_value="linear", allowed_values=["linear", "quadratic"], description="vertical grid transformation"

grid_eta[Float64]=NaN (m), default_value=NaN, description="length scale for vertical grid transformation"

f_porosity[String]="Const", default_value="Const", allowed_values=["Const", "ExpAtten"], description="functional form for porosity vs depth"

zpor[Float64]=0.1 (m), default_value=0.1, description="lengthscale for porosity if f_porosity=ExpAtten"

zdbl[Float64]=0.0004 (m), default_value=0.0004, description="diffusive boundary layer thickness at sediment - water column interface"

w_solute[Bool]=false, default_value=false, description="true to assume w_solute = w_solid at base of column, false to set w_solute=0.0 (ie zero solute velocity at great depth)"

rho_ref[Float64]=1027.0 (kg m-3), default_value=1027.0, description="assumed constant sw density conversion factor"

Methods and Variables for default Parameters

do_sediment_phys

volume (m^3), VT_ReactDependency, description="volume of sediment cells"

volume_total (m^3), VT_ReactDependency, description="total volume of sediment cells"

Abox (m^2), VT_ReactDependency, description="horizontal area of box"

zupper (m), VT_ReactDependency, description="depth of upper surface of box (m) 0 is surface, -100 is depth of 100 m"

zlower (m), VT_ReactDependency, description="depth of lower surface of box (m)"

zmid (m), VT_ReactDependency, description="mean depth of box"

pressure (dbar), VT_ReactDependency, description="sediment pressure"

rho_ref (kg m^-3), VT_ReactDependency, description="density conversion factor"

oceanfloor_Afloor –> oceanfloor.Afloor (m^2), VT_ReactDependency, description="horizontal area of seafloor at sediment surface"

oceanfloor_zfloor –> oceanfloor.zfloor (m), VT_ReactDependency, description="depth of ocean floor (m, -ve)"

oceanfloor_temp –> oceanfloor.temp (K), VT_ReactDependency, description="oceanfloor temperature"

oceanfloor_sal –> oceanfloor.sal (psu), VT_ReactDependency, description="oceanfloor salinity"

oceanfloor_phi –> oceanfloor.phi (), VT_ReactDependency, description="sediment surface porosity"

[oceanfloor_phimin] –> oceanfloor.phimin (), VT_ReactDependency, description="sediment porosity at infinite depth"

oceanfloor_w_accum –> oceanfloor.w_accum (m yr-1), VT_ReactDependency, description="sediment accumulation rate (+ve)"

phi (), VT_ReactProperty, description="porosity (volume fraction of solute phase)"

phi_solid (), VT_ReactProperty, description="1.0 - porosity (volume fraction of solid phase)"

volume_solute (m^3), VT_ReactProperty, description="solute volume of sediment cells"

volume_solid (m^3), VT_ReactProperty, description="solid volume of sediment cells"

temp (Kelvin), VT_ReactProperty, description="sediment temperature"

sal (psu), VT_ReactProperty, description="sediment salinity"

w_solid (m yr-1), VT_ReactProperty, description="solid phase advection velocity (downwards is -ve)"

w_solute (m yr-1), VT_ReactProperty, description="solute phase advection velocity (downwards is -ve)"

Dfac (), VT_ReactProperty, description="tortuoisity-dependent multiplier for solute diffusivity"

PALEOsediment.Sediment.SedimentBioRates.ReactionSedimentBioRates — Type

ReactionSedimentBioRates

Calculate sediment bioturbation and bioirrigation rates (for use by sediment transport).

Literature compilation of functional forms for dependency on oceanfloor oxygen, Corg flux, and depth within sediment, from (Boudreau, 1996), (Archer, 2002), (Arndt et al., 2011), (Dale et al., 2015)

Parameters

f_bioTurbRate[String]="Prescribed", default_value="Prescribed", allowed_values=["Prescribed", "CorgArcher2002"], description="functional form for bioturbation max rate"

f_bioTurbDepth[String]="ConstCutoff", default_value="ConstCutoff", allowed_values=["ConstCutoff", "ErfcCutoff", "Exp1Cutoff", "Exp2Cutoff"], description="functional form of bioturbation rate with depth in sediment"

f_bioIrrigDepth[String]="ConstCutoff", default_value="ConstCutoff", allowed_values=["ConstCutoff", "ErfcCutoff", "Exp1Cutoff", "Exp2Cutoff"], description="functional form of bioirrigation rate with depth in sediment"

f_bioO2[String]="None", default_value="None", allowed_values=["None", "MM", "Dale2015"], description="functional form of bioturbation and bioirrigation sensitivity to oceanfloor oxygen"

bioO2halfmax[Float64]=0.02 (mol m-3), default_value=0.02, description="oceanfloor [O2] for 50% decrease in bioturbation/bioirrigation"

bioO2decreaserate[Float64]=0.012 (mol m-3), default_value=0.012, description="oceanfloor [O2] sharpness of decrease in bioturbation/bioirrigation"

separate_zbio[Bool]=false, default_value=false, description="true to use zbioturb, zbioirrig variables, false to use zbio for both bioturbation and bioirrigation"

Methods and Variables for default Parameters

METHODS PALEOboxes.DocStrings.Methods(:methods_do) exception: ErrorException("PALEOsediment.Sediment.SedimentBioRates.ReactionSedimentBioRates register_methods! not implemented")